==> code/resource_estimate/01_handling_unknown_count_units.Rmd <==
---
title: "Handling unknown count units (for resource estimation)"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Overall goal**

Estimate resource availability per site and year, in terms of calories available to hummingbirds.

**Problem**

Count units were not always recorded during field work, so it is not always clear whether the resource count reflects flowers or groups of flowers (inflorescences).

**Goals of this code**

1. Assign count unit for species where count unit can be reasonably assumed based on either:

(a) related species
(b) plant's morphology (i.e., plants that produce single flowers)
(c) the typical count unit used when it was recorded.

2. For species where count unit is still unknown, create a 'low' and 'high' estimate; 'low' assumes flower, 'high' assumes inflorescence. Inflorescence counts will be multiplied by flowers/inflorescence in later step, yielding higher overall resource counts.

```{r setup, include=FALSE}
library(tidyr)
library(dplyr)
library(ggplot2)
```

## Step 1: Get data

Before this point, I manually assigned (known) count units in Excel if a count unit was indicated in the notes column. When flower data were available in the notes column, I used those data over data at the level of inflorescences or trees.

```{r step1, warning=FALSE}
#Shorten "inflorescence" to "inflor"
#Create column designating known vs. unknown count units (known = count unit explicitly recorded in field)
resource_data <- read.csv("../../data/import/data/resource_counts/Resource_counts_2016-2018.csv") %>%
  mutate(count_unit = ifelse(count_unit == "inflorescence", "inflor", as.character(count_unit)), 
         count_unit_status = ifelse(is.na(count_unit), "unknown", "known"))
```

## Step 2: Deal with unknown species that have a known genus (e.g., COSTUS, PASSIFLORA, HELICONIA) or identifying descriptions

I can still include these species in the calorie estimate by substituting a closely related species. To do this, I will create a new column called 'plant_species_for_calories'.  *Remember, this is just for the resource/calorie estimate! Use original 'plant_species' to calculate other variables, e.g., species richness.*

Which unknown species include genera -- or somewhat familiar/useful descriptions -- and could be reasonably approximated with a known species? Preferably the approximate species will also have nectar data.
```{r step2a, warning=FALSE}
unknown_species <- resource_data %>%
  filter(grepl("UNKN", plant_species)) %>%
  distinct(plant_species)
```

Species substitutions

- UNKN-PASSIFLORA --> PACO (*Passiflora coccinea*)
- UNKN-HELICONIA --> HETO (*Heliconia tortuosa*)
- UNKN-COSTUS --> COLA (*Costus laevis*)
- UNKN-HIBISCUS --> AMAP (*Malvaviscus achanioides*- red amapola)
- UNKN-BROMELIA --> GUZA (*Guzmania zahnii* - yellow bromeliad)
- UNKN-PAV-MOR-PEL --> PAV-MOR (*Dicliptera trifurca*)

Create new column for approximate species. Use this column for rest of calorie estimation
```{r step2b, warning=FALSE}
resource_data02 <- resource_data %>%
  mutate(plant_species_for_calories = case_when(plant_species == "UNKN-PASSIFLORA" ~ "PACO",
                                               plant_species == "UNKN-HELICONIA" ~ "HETO",
                                               plant_species == "UNKN-COSTUS" ~ "COLA",
                                               plant_species == "UNKN-HIBISCUS" ~ "AMAP",
                                               plant_species == "UNKN-BROMELIA" ~ "GUZA", 
                                               plant_species == "UNKN-PAV-MOR-PEL" ~ "PAV-MOR",
                                               TRUE ~ as.character(plant_species)))%>%
  select(year:plant_family, plant_species_for_calories, count_unit, count_unit_status, everything())
```

## Step 3: Assign count units based on knowledge of the plant species/count methods used

- Assign counting unit of 'bract' to Heliconia species (because the project protocol for counting Heliconias was to count the number of open bracts)

- Assign counting unit of 'flower' to species that do not have a countable inflorescence. For the following species, counting them in any way other than **flower** would be very unintuitive, based on their appearance/growing habit:

    * PAV-MOR
    * PAV-ROSA
    * AMAP
    * PACO
    * PAVI
  
- Assign counting unit of 'inflor' to species that do not have countable flowers. For the following species, counting them in any other way than **inflorescence** would be very difficult due to growing habit

    * GUMA (inflorescences - "orange puffs" - grow on a vine, typically out of reach)
  
*Note: need to use 'plant_species_for_calories' so that unknown/approximated species get assigned count units*
```{r step3, warning=FALSE}
#Species with no inflorescence to count
species_counted_as_flowers <- c("PAV-MOR", "PAV-ROSA", "AMAP", "PACO", "PAVI")
species_counted_as_inflor <- c("GUMA")


#Add count units and note about reason
resource_data03 <-resource_data02 %>%
  
  #Assign "bract" to Heliconia species
  mutate(count_unit_source = ifelse(count_unit_status == "unknown" & grepl(pattern = "HE", plant_species_for_calories), "Counting protocol for Heliconias", ""),
         count_unit_status = ifelse(count_unit_status == "unknown" & grepl(pattern = "HE", plant_species_for_calories), "assumed", count_unit_status),
         count_unit = ifelse((is.na(count_unit) | count_unit == "inflor") & grepl(pattern = "HE", plant_species_for_calories), "bract", as.character(count_unit))) %>%
  
  #Assign "flower" to species presumably counted as flowers
  mutate(count_unit_source = ifelse(is.na(count_unit) & (plant_species_for_calories %in% species_counted_as_flowers), "No countable inflorescence", count_unit_source),
         count_unit_status = ifelse(is.na(count_unit) & (plant_species_for_calories %in% species_counted_as_flowers), "assumed", count_unit_status),
         count_unit = ifelse(is.na(count_unit) & (plant_species_for_calories %in% species_counted_as_flowers), "flower", as.character(count_unit))) %>%
  
  #Assign "inflor" to species presumably counted as inflorescences
  mutate(count_unit_source = ifelse(is.na(count_unit) & (plant_species_for_calories %in% species_counted_as_inflor), "No countable flower", count_unit_source),
         count_unit_status = ifelse(is.na(count_unit) & (plant_species_for_calories %in% species_counted_as_inflor), "assumed", count_unit_status),
         count_unit = ifelse(is.na(count_unit) & (plant_species_for_calories == "GUMA"), "inflor", as.character(count_unit))) %>%
  
  #Reorder columns
  select(year:count_unit_status, count_unit_source, everything())
```

## Step 4: Use data about known counting units to estimate unknown counting units

In 2018, count unit was recorded regularly using the Survey213 mobile app. Here, I compare the distribution of 'known' count units to the distributino of 'unknown' count units. If they seem pretty close, I will assumed the known unit. If bimodal distribution, this could indicate multiple count units used.

*Note: Density function needs at least two points, so will only look at species with >2 points*
```{r step4a, warning=FALSE, fig.height=8, fig.width=8}
some_known <- resource_data03 %>%
  group_by(plant_species, plant_scientific_name, count_unit_status) %>%
  summarise(num_rows = n()) %>%
  ungroup() %>%
  arrange(desc(num_rows)) %>%
  pivot_wider(names_from = count_unit_status, values_from = num_rows, values_fill = 0) %>%
  filter(known > 2 & unknown > 2)

#Prepare data for plotting
#1. Remove rows with missing resource counts
#2. Subset data to known species with enough known and unknown counts
#3. Creating new column for count status (adding info about known count unit)
density_plot_data <- resource_data03 %>%
          filter(!is.na(num_other_resources)) %>% 
          filter(plant_species %in% some_known$plant_species) %>% 
          unite(count_info, count_unit_status, count_unit, sep = "-") %>%
          mutate(count_info = ifelse(count_info == "unknown-NA", "unknown", count_info))
  
#Make density plot(s). To spread out X axis, use log(number of resources)
density_plots <- ggplot(data = density_plot_data, aes(x = log(num_other_resources), fill = count_info)) +
            geom_density(alpha = 0.4) +
            facet_wrap(. ~ plant_species, nrow = NULL, ncol = NULL, scales = "free") +
            scale_fill_manual(values = c("blue", "red", "gold", "white")) +
            labs(fill = "Count unit", x = "Ln(number of resources counted)") +
            theme(legend.position = "bottom")
            
            
density_plots
```

**Results of this approach**

Based on overlapping distributions, I am confident about 3 species:

* GINGY = inflorescence
* MARA = inflorescence
* MARA-VER = inflorescence

Add count units for species I am confident about
```{r step4b, warning=FALSE}
resource_data04 <- resource_data03 %>%
  mutate(count_unit_source = ifelse(is.na(count_unit) & grepl(pattern = "GINGY|MARA|MARA-VER", plant_species_for_calories), "Known vs. unknown distributions", count_unit_source),
         count_unit_status = ifelse(is.na(count_unit) & grepl(pattern = "GINGY|MARA|MARA-VER", plant_species_for_calories), "assumed", count_unit_status),
         count_unit = ifelse(is.na(count_unit) & grepl(pattern = "GINGY|MARA|MARA-VER", plant_species_for_calories), "inflor", as.character(count_unit)))
```

## Step 5: For remaining species with unknown count unit, assign a 'low' estimate (i.e., flower) and 'high' estimate (i.e., inflorescence)

**How many species (and plants) still had unknown count units?**
```{r step5a, warning=FALSE}
still_unknown <- resource_data04 %>%
  filter(count_unit_status == "unknown") %>%
  group_by(plant_species, plant_scientific_name, plant_species_for_calories) %>%
  summarise(num_rows = n()) %>%
  ungroup() %>%
  arrange(desc(num_rows)) 

#How many species still have unknown count units? Filtered to species that could be identified to genus
filter(still_unknown, !is.na(plant_scientific_name)) %>% pull(plant_scientific_name) %>% length()

#How many rows (plants) have unknown count units?
unknown_plants <- filter(still_unknown, !is.na(plant_scientific_name)) %>% 
  summarise(num_rows = sum(num_rows)) %>%
  pull(num_rows)

unknown_plants

#What percentage of plants is this?
(total_plants <- nrow(resource_data04))

unknown_plants/total_plants
```

**Assign a high estimate and a low estimate to species where I was not able to figure out the most likely count unit**
```{r step5b, warning=FALSE}
resource_data05 <- resource_data04 %>%
  mutate(high = ifelse(count_unit_status == "unknown", "inflor", as.character(count_unit)),
         low = ifelse(count_unit_status == "unknown", "flower", as.character(count_unit)))
```

*Note: Above here data is in wide form, ~6000 rows. Below here, number of rows will double*

## Step 6: Create columns for # bracts, # flowers, # inflorescences, and # trees

Organizing the data this way will make it easy to convert [bract/inflor/tree] to flowers in a later step; will need to multiply by # flowers/unit

```{r step6, warning=FALSE}
#1. Convert data into long form & make new indicator column ('count_estimate_high_low') for whether the estimate is high or low
resource_data06 <-  resource_data05 %>%
  select(-count_unit) %>% #Will replace this with the data in the 'high' and 'low' column
  pivot_longer(cols = c("high", "low"), names_to = "count_estimate_high_low", values_to = "count_unit") %>%
  select(year:count_unit_source, count_unit, count_estimate_high_low, everything())

#Currently, the data only has columns for # HETO bracts, # HETO flowers, and # resources...with a column ('count_unit') indicating how resources were counted
#1. Make a new column for number of bracts (combining num_heto_bracts and num_other_resources; latter column was used for non-HETO Heliconia species)
#2. Make a new column for number of flowers (combining num_heto_flowers and num_other_resources)
#3. Make a new column for number of inflorescences
#4. Make a new column for number of trees
resource_data07 <- resource_data06 %>%
  mutate(num_bracts_temp = ifelse(count_unit == "bract", num_other_resources, NA)) %>%
  unite(num_bracts, num_heto_bracts, num_bracts_temp, sep = "") %>%
  mutate(num_bracts = gsub("NA", "", num_bracts), 
         num_bracts = ifelse(num_bracts == "", NA, num_bracts)) %>%
  
  mutate(num_other_flowers = ifelse(count_unit == "flower", num_other_resources, NA)) %>%
  unite(num_flowers, num_heto_flowers, num_other_flowers, sep = "") %>%
  mutate(num_flowers = gsub("NA", "", num_flowers), 
         num_flowers = ifelse(num_flowers == "", NA, num_flowers)) %>%

  mutate(num_inflors = ifelse(count_unit == "inflor", num_other_resources, NA),
         num_inflors = gsub("NA", "", num_inflors)) %>%
  
  mutate(num_trees = ifelse(count_unit == "tree", num_other_resources, NA),
         num_trees = gsub("NA", "", num_trees)) %>%
  
  select(year:count_unit, num_bracts, num_flowers, num_inflors, num_trees, num_other_resources, everything()) %>% 
  mutate(count_unit_source = ifelse(count_unit_status == "known", NA, count_unit_source)) %>%
  select(-num_other_resources) #Don't need this column anymore
```

## Step 7: Export data
```{r step7}
write.csv(resource_data07, "../../data/export/intermediate/Resource_counts_2016-2018_with_count_units.csv", row.names = FALSE) 
```

*Next steps: estimating flowers-per-inflorescence and flowers-per-tree, then estimating calories*
