==> code/resource_estimate/04_calculating_calories_per_flower.Rmd <==
---
title: "Summarizing nectar measurements (for resource estimation)"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Overall goal**

Estimate resource availability per site and year, in terms of calories available to hummingbirds.

**Goals of this code**

Calculate the number of calories per flower, based on nectar volume and concentration.

Steps involved in this calculation (following Bolton *et al.* 1987, Kearns & Inouye 1993)

1. Summarize mean volume and concentration (note: did this in previous step)
2. Use Kearns & Inouye 1993 (Table 5-2) to convert Brix (g/g) to g/L. To use this table, round nectarConc (Brix %, g/g) to nearest 0.5 (if 10 or less), 1.0 (if >10), or 2.0 (if >20)
3. Convert nectar volume (in microliters) to liters
4.  Multiply #2 (g/L) by #3 (L) by 3.94 kilocalories/g

*Oof my naming conventions are V E R Y wonky here, due to a very brief foray into camelCase years ago. Leaving them be for now, since the code works fine.*

```{r setup, include=FALSE}
library(dplyr)
library(tidyr)
```

## Step 1: Import data
```{r step1a, message = FALSE}
#Summary of nectar volume/concentration measurements collected in field
#Need to rename GIN-ROS-AMA because there is another unknown Renealmia species
#Decided to use mean nectar volume and nectar concentration (instead of median) since that is value most commonly found in literature
nectar_field <- read.csv("../../data/export/supp_tables/extra/Nectar_volume_and_concentration_summary.csv") %>%
  mutate(plant_scientific_name = ifelse(plant_species == "GIN-ROS-AMA", "Renealmia sp. 2", as.character(plant_scientific_name))) %>%
  select(plant_species, plant_scientific_name, plant_family, nectar_vol = mean_vol, nectar_conc = mean_conc, everything()) %>%
  mutate(data_source = "field_data")

path_name <- c("../../data/import/data/nectar_measurements")

#Kearns & Inouye 1993, Table 5-2 (converts Brix w/w to w/volume)
conversion_table <- read.csv(file.path(path_name, "Kearns_Inouye_1993_Table_5-2.csv")) %>%
  rename(nectar_conc_approx = nectarConc_approx, nectar_conc_gL = nectarConc_gL)

#Data from literature, for species that were missing field data. There are still blanks for some species; I will substitute field data from congeneric species in next step
nectar_lit <- read.csv(file.path(path_name, "Nectar_measurements_from_literature_20200528.csv")) %>%
  select(plant_species, plant_scientific_name, plant_scientific_name_sub, plant_family, nectar_vol, nectar_conc, lit_site = site, lit_sources) %>%
  arrange(lit_sources)
```

## Step 2: Get data ready for calorie calculations

For species that are still missing nectar volume/concentration data, fill in with the nectar measurements from a congeneric species (data we collected in the field). Also combine our nectar measurements with data from literature
```{r step2, warning = FALSE, message = FALSE}
#plant_scientific_name_sub is the substitute (congeneric) plant species whose data I will use
values_for_missing_species <- nectar_field %>%
  select(plant_scientific_name_sub = plant_scientific_name, volume_this_study = nectar_vol, conc_this_study = nectar_conc)

nectar_missing_species <- nectar_lit %>%
  filter(is.na(nectar_vol) | is.na(nectar_conc)) %>%
  left_join(values_for_missing_species) %>%
  mutate(nectar_vol = ifelse(is.na(nectar_vol), volume_this_study, nectar_vol)) %>%
  mutate(nectar_conc = ifelse(is.na(nectar_conc), conc_this_study, nectar_conc)) %>%
  select(-contains("this_study")) %>%
  mutate(data_source = "field_data_subbed")

nectar_data <- nectar_lit %>%
  filter(!is.na(nectar_vol)) %>%
  mutate(data_source = "literature") %>%
  bind_rows(nectar_field) %>%
  bind_rows(nectar_missing_species) %>%
  select(plant_species, plant_scientific_name, plant_scientific_name_sub, everything())

#Check: are any species missing data?
check01 <- nectar_data %>%
  filter(is.na(nectar_vol) | is.na(nectar_conc))
```

## Step 3: Calculate calories
```{r step3, message = FALSE}
#1. Round nectar concentrations (Brix %) to match what is in Kearns & Inouye table
#2. Join to Kearns and Inouye table. New column is grams of sugar per liter nectar (nectarConc_gL)
#3. Convert nectar volume to L (from microliters)
#4. Calculate calories per flower
nectar_data_with_calories <- nectar_data %>%
  mutate(nectar_conc_approx = case_when(nectar_conc <= 10 ~ plyr::round_any(nectar_conc,accuracy = 0.5),
                                        nectar_conc > 10 & nectar_conc <= 20 ~ plyr::round_any(nectar_conc,accuracy = 1.0),
                                        nectar_conc > 20 ~ plyr::round_any(nectar_conc,accuracy = 2.0))) %>%
  left_join(conversion_table) %>%
  mutate(nectar_vol_L = nectar_vol/1000/1000) %>% #Microliters to milliliters to liters
  mutate(kilocalories = nectar_conc_gL * nectar_vol_L * 3.94) %>% #Sucrose provides 3.94 kilocalories/g
  mutate(calories = kilocalories * 1000) %>%
  mutate(calories = round(calories, digits = 2)) %>%
  mutate(kilocalories = round(kilocalories, digits = 5)) %>%
  select(-nectar_conc_approx, -nectar_conc_gL, -nectar_vol_L) %>%
  select(data_source, plant_family, plant_species, plant_scientific_name, plant_scientific_name_sub, mean_vol = nectar_vol, sd_vol, contains("vol"), mean_conc = nectar_conc, sd_conc, contains("conc"), calories, kilocalories, starts_with("numFlowers"), everything()) %>%
  arrange(data_source, plant_family, plant_scientific_name) 
```

## Step 4: Export data
```{r step4, message = FALSE}
path_name <- c("../../data/export/supp_tables/extra")

#extra tables - not including in final Supporting information
write.csv(nectar_data_with_calories, file.path(path_name, "Nectar_volume_concentration_and_calories_summary.csv"), row.names = FALSE)
```
