==> code/resource_estimate/05_calculating_calories_per_plant.Rmd <==
---
title: "Calculating calories per plant (for resource estimation)"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Overall goal**

Estimate resource availability per site and year, in terms of calories available to hummingbirds.

**Goal of this code**

Calculate the number of calories per plant (row in resource count dataset) using flowers-per-inflorescence keys and calorie info calculated in previous steps. 

Steps

1. Import plant counts (version with count units) from 2016-2018
2. Import key for HELICONIA FLOWERS PER BRACT
3. Import key for FLOWERS per INFLORESCENCE (or TREE)
4. Import nectar data (CALORIES per FLOWER)
5. Convert to flowers, if data not already in flowers ('num_flowers_estimate')
        
    - If species is HETO, multiply by appropriate number of open flowers per open number of bracts
    - If unit == inflor, multiply by 'per_inflor'
    - If unit == tree, multiply by 'per_tree'
    
6. Calculate CALORIES per PLANT

    - Multiply 'num_flowers_estimate' by calories per flower
    
Any unknown species that could be identified to genus have already been assigned a 'plant_species_for_calories', which is a related (congeneric) species for the purpose of estimating flowers & calories. This is the column to join by.

```{r setup, include=FALSE}
library(tidyr)
library(dplyr)
```
  
## Step 1: Import data
```{r step1, warning = FALSE, message = FALSE}
path_name <- c("../../data/export/intermediate")

#Resource data
resource_counts <- read.csv(file.path(path_name, "Resource_counts_2016-2018_with_count_units.csv")) %>%
  select(-row_id)

path_name <- c("../../data/export/supp_tables/extra")

#Heliconia flowers/bract. I.e. for a given number of open bracts, how many flowers are probably open?
flowers_per_bracts <- read.csv(file.path(path_name, "Heliconia_flowers_per_bracts_key.csv")) %>%
  mutate(plant_species_for_calories = "HETO") %>% 
  select(plant_species_for_calories, num_bracts, flowers_per)

#Flowers-per-inflor (/tree) key for non-Heliconia species
flowers_per_inflor <- read.csv(file.path(path_name, "Non-Heliconia_flowers_per_inflorescence_key.csv")) %>% 
  select(plant_species_for_calories = plant_species, count_unit, flowers_per)

#Nectar data with info about calories per flower
calories_per_flower <- read.csv(file.path(path_name, "Nectar_volume_concentration_and_calories_summary.csv")) %>%
    select(plant_species_for_calories = plant_species, calories)
```

## Step 2: Convert number of bracts/inflorescences/trees into number of flowers
```{r step2, warning=FALSE, message = FALSE}
#1. Assign HETO as the 'plant_species_for_calories' for ALL Heliconia species, not just tortuosa.
#2. Add imported data
resource_counts02 <- resource_counts %>%
  mutate(plant_species_for_calories = ifelse(grepl("Heliconia", plant_scientific_name), "HETO", as.character(plant_species_for_calories))) %>%
  left_join(flowers_per_bracts, by = c("plant_species_for_calories", "num_bracts")) %>%
  left_join(flowers_per_inflor, by = c("plant_species_for_calories", "count_unit")) %>%
  left_join(calories_per_flower, by = c("plant_species_for_calories")) %>%
  unite(flowers_per, flowers_per.x, flowers_per.y, sep = "") %>%
  mutate(flowers_per = gsub("NA", "", flowers_per, fixed = TRUE))

#Check: what rows do not have any flowers-per (bract/infor/tree) data? Heliconia species with no bract data (either no data recorded, or only data about number of flowers recorded)
check01 <- resource_counts02 %>%
  filter(count_unit != "flower") %>%
  filter(!grepl("UNKN", plant_species)) %>%
  filter(is.na(flowers_per) | flowers_per == "")

#Convert bracts/inflor/trees to number of flowers ('num_flowers_estimate')
#1. Keep flower data if it is available...
#2. ...except for Heliconia species (then use bracts-to-flower key for consistency across years)
#3. ...except if there is not bract info (then use # flowers)
resource_counts03 <- resource_counts02 %>%
  mutate(flowers_per = as.numeric(flowers_per)) %>%
  mutate(num_flowers_estimate = ifelse(!is.na(num_flowers), num_flowers, NA),  
         num_flowers_estimate =ifelse(plant_species_for_calories == "HETO", flowers_per, num_flowers_estimate), 
         num_flowers_estimate = ifelse(plant_species_for_calories=="HETO" & is.na(num_bracts), num_flowers, num_flowers_estimate), 
         num_flowers_estimate = ifelse((is.na(num_flowers) & count_unit=="inflor"), num_inflors*flowers_per, num_flowers_estimate),
         num_flowers_estimate = ifelse((is.na(num_flowers) & count_unit=="tree"), num_trees*flowers_per, num_flowers_estimate))

#Check: is the math correct?
check02 <- resource_counts03 %>%
  select(count_estimate_high_low, plant_species, plant_species_for_calories, count_unit, num_bracts, num_flowers, num_inflors, num_trees, flowers_per, num_flowers_estimate, notes_data)

#Check: how many rows do not have a num_flowers_estimate? Why? 
check03 <- resource_counts03 %>%
  filter(is.na(num_flowers_estimate))

#Answer: When count is NA. This includes situations species is 'unknown' and count_unit_estimate='high'/count_unit='inflor'. Latter happens because I do not have flowers-per-inflor data for unknown species. If want to compare high/low estimate summaries, should remove unknown species

#Check: how many rows do not have num_flowers_estimate? Why? (Aside from being an unknown species)
check04 <- resource_counts03 %>%
  filter(is.na(num_flowers_estimate)) %>%
  filter(!grepl("UNKN", plant_species))

#Answer: When count data are NA. For example, when rows were assumed to be duplicate counts (e.g., p137 MARIPA and p10 PORO), or when data are just missing.
```

Possible solutions for plants (rows) missing count data:

A) Add a minimum number of flowers/inflorescences (i.e. 1)

B) Add median, mean, or most common number of flowers for that plant species (if data available). And if those data are not available, add the minimum. 

*AS OF 5/27/2020, DECIDED NOT TO WORRY ABOUT THIS. ONLY ~25 PLANTS ARE MISSING COUNT INFO*

## Step 3: Calculate calories per plant by multiplying number of flowers by calories per flower

```{r step3, message = FALSE}
#This is calories, not kilocalories
calories_per_plant <- resource_counts03 %>%
  mutate(calories_per_plant = as.numeric(num_flowers_estimate)*as.numeric(calories)) %>%
  select(count_estimate_high_low, everything()) %>%
  arrange(count_estimate_high_low)

#Check: are there any species missing nectar data?
nectar_species <- calories_per_flower %>% distinct(plant_species_for_calories)

check05 <- resource_counts03 %>% 
  filter(!(plant_species_for_calories %in% nectar_species$plant_species_for_calories))

#Answer: only random unknown species (not known to genus)

#Check: which rows have zero calories? And why?
check06 <- calories_per_plant %>%
  filter(calories_per_plant == 0)

#Answer: species with 0 (calculated) flowers. This is mostly Heliconia with 0-2 open bracts.

#Check: which rows have NA calories? And why? Did not have any 'num_flowers_estimate'. Same as check04
check08 <- calories_per_plant %>%
  filter(is.na(calories_per_plant)) %>%
  filter(!grepl("UNKN", plant_species))
```

## Step 4: Export data (will be summarized to patch level in later step)

```{r step4, message = FALSE}
path_name <- c("../../data/export/intermediate")
write.csv(calories_per_plant, file.path(path_name, "Resource_counts_2016-2018_with_calories.csv"), row.names = FALSE)
```
