==> code/resource_estimate/06_summarizing_resource_availability.Rmd <==
---
title: "Summarizing resource availability"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goal**

Summarize resource availability data in several ways:

1. *Heliconia* removed in all sites. This is the quantitative version of treatement, as controls will have zero *Heliconia* removed.

2. Percentage of *Heliconia* removed per treatment area

- That is, *Heliconia* removed divided by all *Heliconia* available (*Heliconia* removed + *Heliconia* NOT removed). This % should be pretty high, since we covered all plants except two focal plants the occasional *Heliconia* plants that were inaccessible.

3. *Heliconia* density. This is just to get an idea for natural densities of Heliconia, across all sites and years.

4. Percentage of resources removed per treatment area

- That is, *Heliconia* removed divided by all resources available (*Heliconia* removed + *Heliconia* NOT removed + Non-*Heliconia*).

**Approach**

Calories is the most biologically relevant unit of measurement for hummingbirds, so that unit will be my main focus.

However, I will also summarize the data in terms of plants and flowers, since there might be situations where these units are more intuitive to discuss. For instance, I'd like to talk about the percentage of *Heliconia* plants removed (well, technically inflorescences, because each *Heliconia* row is an inflorescence).

Also, I will tailor the resource estimates to different hummingbird species (or group of species), because hummingbirds do not visit each plant species equally. I will use the camera data to determine which plant species are visited by different hummingbird species, and how frequently they are visited. This tailoring will lead to several levels of calorie estimates, in order of increasing complexity (and presumably biological relevance):

- No tailoring ("base estimate" of % calories removed): includes all plant species 

- Subset ("low estimate" of % calories removed): only includes plant species with confirmed visitation by hummingbirds (or groups of hummingbirds species) 

- Weighted ("high estimate" of % calories removed): only includes plant species with confirmed visitation by hummingbirds (or groups of hummingbirds species), but also weights the value of each plant species by the relative hummingbird visitation rate to that species. If the plant species is not visited at all, it receives a weight of zero; this is the same as the subsetting step above. Plants with non-zero sighting rates receive a weight between 0 and 1 (weight of 1 = most frequently visited species).

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
library(dplyr)
library(tidyr)
library(ggplot2)
library(purrr)
library(ggpubr)
```

## Step 1: Import data and helper functions
```{r step1}
#Resource survey data. Use this unfiltered dataset for any calculations of species richness. Note that there are 'low' and 'high' estimates based on counting units.
resource_data <- read.csv("../../data/export/intermediate/Resource_counts_2016-2018_with_calories.csv") %>% 
  select(-calories, -flowers_per) %>%
  select(year, patch, everything())

#Focal area sizes
focal_area_sizes <- read.csv("../../data/export/intermediate/Focal_area_sizes_and_centroids.csv") %>%
  select(year, patch, control_treatment, area)

#Elevation
elevation <- readxl::read_excel("../../data/import/data/Site_elevations.xlsx")

#Plant scientific names
plant_names <- read.csv("../../data/import/data/Plant_codes_species_list.csv") %>%
  select(plant_species, plant_scientific_name, plant_family) %>%
  filter(!is.na(plant_species) & plant_species != "")

#Camera data reflected natural visitation, but removing night-blooming species. These will not be in the calorie estimate, so I don't need to include them
camera_data <- read.csv("../../data/export/intermediate/Camera_data_filtered_for_normal_visitation.csv") %>%
  filter(plant_species != "OJO" & plant_species != "CEPU")

#Functions
source("../../code/helper_functions/Calculate_basic_summary_stats.R")
source("../../code/helper_functions/Summarize_camera_data.R")
source("../../code/helper_functions/Make_barplot.R")
```

## Step 2: Make list of all plant species detected, for Supp methods, then filter data for resource estimate

```{r step2a}
#How many species in each site?
species_per_site <- resource_data %>%
  distinct(patch, year, plant_family, plant_scientific_name, plant_species) %>%
  group_by(patch, year) %>%
  summarise(num_species = n()) %>%
  ungroup()

#What was mean/median number of species per site?
species_per_site_sum <- species_per_site %>%
  calculate_basic_summary_stats(variable = num_species)

species_per_site_sum

#How many species total?
species_list <- resource_data %>%
  distinct(plant_family, plant_scientific_name, plant_species, plant_species_for_calories) %>%
  mutate(known_vs_unknown = ifelse(grepl("UNKN", plant_species), "unknown", "known")) %>%
  arrange(known_vs_unknown)

#How many known vs unknown?
species_list_sum <- species_list %>%
  group_by(known_vs_unknown) %>%
  summarise(num_species = n()) %>%
  ungroup()

species_list_sum

#Unknown species that did not have an approximate (i.e., genus level) species substituted 
completely_unknown_species <- species_list %>%
  filter(grepl("UNKN", plant_species_for_calories))

nrow(completely_unknown_species)

#How many plants was this?
completely_unknown_plants <- resource_data %>%
  filter(count_estimate_high_low == "high") %>%
  filter(grepl("UNKN", plant_species_for_calories))

nrow(completely_unknown_plants)
```

Remove unknown species (i.e.,, UNKN-_____)

  - For unknown species, I can calculate number of plants (rows) but not necessarily the numbers of flowers and will never be able to calculate the numbers of calories.
        
  - Situations when I can/cannot get number of flowers from unknown species:
    * Can include unknown if count unit = flower. This occurs if count unit is known (and is 'flower') or in the 'low' flower estimate, which assumes flower for all unknown count units
    * Cannot include unknowns in 'high' estimate, because do not know flowers per inflorescence.
      
  - But for high & low plots of flower abundance to be comparable, need to remove unknowns from both, so I will remove.

Remove night-blooming species (i.e., OJO, CEPU) 

  - Bat-pollinated species produce nectar at night. Though some nectar may be leftover and available to hummingbirds at dawn/dusk, I don't know how much that would generally be.
  
```{r step2b}
#Resource data. Use this filtered dataste for calculations of abundance and calories.
resource_data_filtered <- resource_data %>%
  filter(!grepl("UNKN", plant_species_for_calories)) %>%  #Remove unknown species that did not have an approximate (i.e., genus level) species substituted
  filter(plant_species != "OJO" & plant_species != "CEPU") #Remove species that open at night and are visited by bats
```

## Step 3: Add additional columns that will help create summaries of interest

There is currently a column called 'covered', but it only applies to *Heliconia* (i.e., *Heliconia tortuosa*, aka HETO) in treatment sites; values are NA for control sites and any non-HETO species.

To make cover status more explicit, here I add two columns reflecting all species in all sites:

1. removed_remaining: indicates whether resource is covered (removed) or not covered (remaining)

2. heto_nonhento: indicates whether species is Heliconia (HETO) or not (non-HETO)

Various combinations of these can be used in summaries.

```{r step3}
#Add categories to resource data
resource_data_to_summarize <- resource_data_filtered %>%
  mutate(removed_remaining = case_when(covered == "Y" ~ "removed",
                                       covered == "N" ~ "remaining",
                                       covered == "NA" | is.na(covered) ~ "remaining",
                                       TRUE ~ as.character(NA))) %>% 
  mutate(heto_nonheto = case_when(plant_species == "HETO" ~ "heto_only",
                                  plant_species != "HETO" ~ "nonheto",
                                  TRUE ~ as.character(NA))) %>%
  select(-covered)

#Add additional categories to focal area data, which has all replicates - including 2016 controls. This will allow me to left_join these data to summaries without having NAs for count unit estimate, etc.
focal_area_sizes_to_join <- focal_area_sizes %>%
  crossing(tibble(count_estimate_high_low = c("high", "low"))) %>%
  crossing(tibble(heto_nonheto = c("heto_only", "nonheto")))
```

## Step 4: Calculate resources removed and resources remaining
```{r step4}
#At level of CALORIES
resource_data_sum_calories <- resource_data_to_summarize %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_calories = sum(calories_per_plant, na.rm = TRUE)) %>%
  ungroup()

#At level of PLANTS, FLOWERS, and CALORIES for Heliconia
heto_inflor_removed_remaining <- resource_data_to_summarize %>%
  filter(plant_species == "HETO" & count_estimate_high_low == "high") %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_plants = n(),
            num_flowers = sum(num_flowers_estimate, na.rm = TRUE),
            num_calories = sum(calories_per_plant, na.rm = TRUE)) %>%
  ungroup()
```

## Step 5: Calculate *Heliconia* removed in all sites and years (quantitative version of treatment)

Will calculate at multiple levels for Supp material, but will use calories in analysis
```{r step5}
#Note: do not need multiple count estimates here, because count unit was always known (bracts, converted to flowers)
heto_inflor_removed <- heto_inflor_removed_remaining %>%
  filter(removed_remaining == "removed")

#How many inflorescences with open bracts? 
heto_inflor_open_removed <- resource_data_to_summarize %>%
  filter(plant_species == "HETO" & removed_remaining == "removed" & count_estimate_high_low == "high") %>%
  filter(num_bracts > 0) %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_plants_open = n()) %>%
  ungroup()

#How many with at least 3 open bracts?
heto_inflor_gt3_removed <- resource_data_to_summarize %>%
  filter(plant_species == "HETO" & removed_remaining == "removed" & count_estimate_high_low == "high") %>%
  filter(num_bracts >= 3) %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_plants_gt3 = n()) %>%
  ungroup()

#How many flowers?
heto_flowers_removed <- resource_data_to_summarize %>%
  filter(plant_species == "HETO" & removed_remaining == "removed" & count_estimate_high_low == "high") %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_flowers = sum(num_flowers_estimate, na.rm = TRUE)) %>%
  ungroup()

#How many calories?
heto_calories_removed <- resource_data_sum_calories %>%
  filter(heto_nonheto == "heto_only") %>%
  filter(count_estimate_high_low == "high") %>%
  pivot_wider(names_from = removed_remaining, values_from = num_calories) %>%
  rename(num_calories = removed)
 
#Combine into table for Supp material
heto_removed <- heto_inflor_removed %>%
  left_join(heto_inflor_open_removed) %>%
  left_join(heto_inflor_gt3_removed) %>%
  left_join(heto_flowers_removed) %>%
  left_join(heto_calories_removed) %>%
  left_join(focal_area_sizes) %>%
  mutate(num_calories_per_ha = num_calories/area) %>%
  select(-heto_nonheto, -count_estimate_high_low, -removed_remaining, -remaining) %>%
  arrange(num_calories_per_ha) %>%
  select(year, patch, contains("plants"), contains("flowers"), contains("calories"), area, everything()) %>%
  mutate(across(where(is.numeric) & !c(area), round))

#With focal area values rounded
heto_removed_supp <- heto_removed %>%
  mutate(across(c(area), ~round(., digits = 2)))

#Summary of Heliconia removed (and focal area sizes)
heto_removed_sum <- heto_removed_supp %>%
  pivot_longer(cols = c(num_plants:num_calories_per_ha, area), names_to = "unit", values_to = "value") %>%
  group_by(unit) %>%
  calculate_basic_summary_stats(variable = value)

heto_removed_sum

#Add zeroes for control replicates (control replicate = zero Heliconia removed)
heto_removed_as_treatment <- focal_area_sizes %>%
  filter(control_treatment == "control") %>%
  bind_rows(heto_removed) %>%
  select(year, patch, control_treatment, area, calories_removed_per_ha = num_calories_per_ha) %>%
  mutate(across(where(anyNA), ~ replace_na(., 0)))
```

## Step 6: Calculate percentage of *Heliconia* removed in treatment areas
```{r step6}
#Do not need multiple count estimates here, because count unit was always known (bracts, converted to flowers)
#If number of HETO plants NA or <2, we must not have recorded the 2 focal plants. Adding these here.
percentage_heto_removed <- heto_inflor_removed_remaining %>%
  pivot_wider(names_from = removed_remaining, values_from = num_plants) %>%
  mutate(remaining = ifelse(remaining < 2 | is.na(remaining), 2, remaining)) %>%
  mutate(percentage_removed = (removed/(removed + remaining)) * 100) %>%
  select(-heto_nonheto, -count_estimate_high_low)

percentage_heto_removed_sum <- percentage_heto_removed %>%
  calculate_basic_summary_stats(variable = percentage_removed)

percentage_heto_removed_sum
```

## Step 7: Calculate natural density of *Heliconia* in all sites
```{r step7}
#Heliconia density
#Summarize across removed and remaining, since I am interested all Heliconia plants (including those that were removed as part of treatment)
heto_densities <- heto_inflor_removed_remaining %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, heto_nonheto) %>%
  summarise(num_plants = sum(num_plants),
            num_flowers = sum(num_flowers),
            num_calories = sum(num_calories)) %>%
  ungroup() %>%
  filter(heto_nonheto == "heto_only") %>%
  filter(count_estimate_high_low == "high") %>%
  left_join(focal_area_sizes_to_join) %>%
  mutate(heto_inflors_per_ha = num_plants/area,
         heto_flowers_per_ha = num_flowers/area,
         heto_calories_per_ha = num_calories/area) %>%
  mutate(across(c(contains("per_ha"), "num_calories"), ~round(., digits = 0))) %>%
  select(year, patch, control_treatment, contains("per_ha")) %>%
  arrange(patch, year)

heto_densities_sum <- heto_densities %>%
  select(-contains("num")) %>%
  pivot_longer(cols = c(heto_inflors_per_ha, heto_flowers_per_ha, heto_calories_per_ha), values_to = "value", names_to = "unit") %>%
  group_by(unit) %>%
  calculate_basic_summary_stats(variable = c(value)) %>%
  mutate(across(where(is.numeric), ~round(., digits = 0)))

heto_densities_sum
```

## Step 8: Calculate percentage of calories removed in treatment areas
```{r step8}
#Summarize across species (HETO and non-HETO), since I am interested in percentage of calories removed out of all calories (including all species)
percentage_calories_removed <- resource_data_sum_calories %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, removed_remaining) %>%
  summarise(num_calories = sum(num_calories)) %>%
  ungroup() %>%
  pivot_wider(names_from = removed_remaining, values_from = num_calories) %>%
  filter(control_treatment == "treatment") %>%
  mutate(percentage_removed = (removed/(removed + remaining)) * 100) %>%
  mutate(across(where(is.numeric), ~round(., digits = 1))) %>%
  arrange(count_estimate_high_low, percentage_removed)

percentage_calories_removed_sum <- percentage_calories_removed %>%
  group_by(count_estimate_high_low) %>%
  calculate_basic_summary_stats(variable = percentage_removed) %>%
  mutate(across(where(is.numeric), ~round(., digits = 1)))

percentage_calories_removed_sum

#This summary has not yet been tailored to any hummingbird species! See steps below...
```

## Step 9: Tailor estimates to hummingbird species groups, part I: calculate weights

Weights are the relative visitation rates of hummingbirds (or groups of hummingbird species) to different plant species.

How much effort went into determining resource weights? Summarize effort for table of weights.
```{r step9a}
#Effort per individual plant species (across sites)
#Note: if >1 plant species in frame, that effort is counted more than once. This leads to higher effort estimates than other effort summaries, which are summarized to level of camera.
effort_by_plant_species = camera_data %>%
  distinct(year, patch, control_treatment, camera_num, plant_species, date_video, exp_phase, file_id, video_length) %>%
  group_by(year, patch, control_treatment, camera_num, plant_species, date_video, exp_phase) %>%
  summarise(video_hours = sum(video_length)) %>%
  ungroup()

#How many hours per plant?
hours_per_plant_species = effort_by_plant_species %>%
  group_by(plant_species) %>%
  summarise(video_hours = sum(video_hours)) %>%
  ungroup()

#How many cameras per plant species?
cameras_per_plant_species <- effort_by_plant_species %>%
  select(year, patch, camera_num, plant_species) %>%
  distinct() %>%
  group_by(plant_species) %>%
  summarise(num_cameras = n())

#How many sites (patches) went into determining the weights? Patches per plant species.
sites_per_plant_species <- effort_by_plant_species %>%
  select(patch, plant_species) %>%
  distinct() %>%
  group_by(plant_species) %>%
  summarise(num_sites = n())

#Combining summaries
effort_for_weights <- cameras_per_plant_species %>%
  left_join(sites_per_plant_species) %>%
  left_join(hours_per_plant_species)
```

Calculate visitation rates and convert to weights

Will calculate for:

- All species seen on camera
- Heliconia specialists: green hermits (GREH) and violet sabrewings (VISA)
- Any other bird groups of interest, which can be added as needed. In previous versions, I included subsets for telemetry species, etc.

Note: visitation rates are called "sighting rates" in the summarizing script; technically the # of sightings in which a hummingbird was confirmed to visit a flower

```{r step9b}
weights_all_species <- calculate_sighting_rates(data = camera_data, level_org = "plant_species_across_sites", level_bird = "camera_spp", sightings = "with_visit", marked = "all", include_unknown_spp = FALSE) %>%
  rates_to_weights(level_bird = "camera_spp") %>%
  mutate(bird_group = "all_spp")

weights_greh_visa <- calculate_sighting_rates(data = camera_data, level_org = "plant_species_across_sites", level_bird = list(greh_visa = c("GREH", "VISA")), sightings = "with_visit", marked = "all", include_unknown_spp = FALSE) %>%
  rates_to_weights(level_bird = list(greh_visa = c("GREH", "VISA")))

#Combine weights
weights <- weights_all_species %>%
  bind_rows(weights_greh_visa) %>%
  left_join(plant_names) %>%
  select(bird_group, plant_species, hours, sightings, sightings_per_hour, max_rate, weight, rank, plant_scientific_name, plant_family)

#Format for Supp table
weights_supp <- weights %>%
  select(-c(max_rate, rank, hours)) %>%
  pivot_wider(names_from = bird_group, values_from = c(sightings, sightings_per_hour, weight), names_glue = "{bird_group}_{.value}") %>%
  arrange(desc(all_spp_weight)) %>%
  left_join(effort_for_weights) %>%
  mutate_if(is.numeric, ~round(., digits = 2)) %>%
  mutate(video_hours = round(video_hours, digits = 0)) %>%
  select(plant_family, plant_species, plant_scientific_name, video_hours, num_cameras, num_sites, starts_with("all_spp"), contains("greh_visa"), everything())
```

## Step 10: Tailor estimates to hummingbird species groups, part II: apply weights to resource data

Want two versions of the calorie estimate, and ultimately % calories removed:

- a SUBSET version, which only includes plant species with confirmed hummingbird visitation (by all species or by GREH/VISA)
- a WEIGHTED version, which is the SUBSET version *plus* weighting based on visitation frequency

```{r step10}
#Weights for calorie estimate
weights_to_join <- weights %>%
  select(bird_group, plant_species_for_calories = plant_species, contains("weight")) %>%
  pivot_wider(names_from = bird_group, values_from = weight, names_glue = "{bird_group}")

#Add weights
resource_data_to_summarize_with_weights <- resource_data_to_summarize %>%
  left_join(weights_to_join) %>%
  pivot_longer(cols = c("all_spp", "greh_visa"), values_to = "weight", names_to = "bird_group") %>%
  mutate(weighted_calories_per_plant = calories_per_plant * weight) %>% #Create new column: calories multiplied by weight. THIS IS THE WEIGHTING STEP.
  select(bird_group, count_estimate_high_low, everything()) %>%
  arrange(bird_group, count_estimate_high_low)

#Summarize for SUBSET
resource_data_sum_calories_subset <- resource_data_to_summarize_with_weights %>%
  filter(weight > 0) %>% #Subset (filter) to plant species that had sightings on camera. THIS IS THE SUBSETTING STEP
  group_by(bird_group, count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_calories = sum(calories_per_plant, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(estimate_name = "low_estimate")

#Summarize for WEIGHTED
resource_data_sum_calories_weighted <- resource_data_to_summarize_with_weights %>%
  group_by(bird_group, count_estimate_high_low, year, patch, control_treatment, removed_remaining, heto_nonheto) %>%
  summarise(num_calories = sum(weighted_calories_per_plant, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(estimate_name = "high_estimate")
```

## Step 11: Tailor estimates to hummingbird species groups, part III: calculate percentage calories removed using the tailored estimates

Combine all estimates + calculate percentage of calories removed
```{r step11}
percentage_calories_removed_tailored <- resource_data_sum_calories %>%
  mutate(estimate_name = "base_estimate") %>%
  mutate(bird_group = "all_spp") %>%
  bind_rows(resource_data_sum_calories_subset) %>%
  bind_rows(resource_data_sum_calories_weighted) %>%
  group_by(estimate_name, count_estimate_high_low, bird_group, year, patch, control_treatment, removed_remaining) %>%
  summarise(num_calories = sum(num_calories)) %>%
  ungroup() %>%
  pivot_wider(names_from = removed_remaining, values_from = num_calories) %>%
  rename(heto_calories_removed = removed, heto_nonheto_calories_remaining = remaining) %>%
  mutate(percentage_removed = (heto_calories_removed/(heto_calories_removed + heto_nonheto_calories_remaining)) * 100) %>%
  mutate(across(c(percentage_removed), ~round(., digits = 1))) %>%
  filter(count_estimate_high_low == "high") %>% #i.e., assuming unknown count unit of inflorescence vs. flower
  select(estimate_name:control_treatment, heto_calories_removed, heto_nonheto_calories_remaining, percentage_removed, -count_estimate_high_low) %>%
  arrange(control_treatment)

#Summary stats
percentage_calories_removed_tailored_sum <- percentage_calories_removed_tailored %>%
  group_by(estimate_name, bird_group) %>%
  calculate_basic_summary_stats(variable = percentage_removed) %>%
  mutate(across(where(is.numeric), ~round(., digits = 1)))

percentage_calories_removed_tailored_sum

#Format for Supp table
percentage_calories_removed_tailored_supp <- percentage_calories_removed_tailored %>%
  select(estimate_name, bird_group, year, patch, control_treatment, percentage_removed) %>% 
  pivot_wider(names_from = c(estimate_name, bird_group), values_from = percentage_removed) %>%
  arrange("base_estimate_all_spp")
```

## Step 12: Calculate background resource availability (i.e., density of non-*Heliconia* resources for each hummingbird group)
```{r step12}
#These data are not actually used in this paper, but want to go ahead and make them available for another analysis
nonheto_densities_tailored <- resource_data_sum_calories %>%
  mutate(estimate_name = "base_estimate") %>%
  mutate(bird_group = "all_spp") %>%
  bind_rows(resource_data_sum_calories_subset) %>%
  bind_rows(resource_data_sum_calories_weighted) %>%
  group_by(estimate_name, count_estimate_high_low, bird_group, year, patch, control_treatment, heto_nonheto) %>%
  summarise(num_calories = sum(num_calories)) %>%
  ungroup() %>%
  pivot_wider(names_from = heto_nonheto, values_from = num_calories) %>%
  rename(nonheto_calories = nonheto) %>%
  filter(count_estimate_high_low == "high") %>% #i.e., assuming unknown count unit of inflorescence vs. flower
  left_join(focal_area_sizes) %>%
  mutate(nonheto_calories_per_ha = nonheto_calories/area) %>%
  select(-count_estimate_high_low, -heto_only, -area)

#Add background resource availability to percentage calories removed dataset
percentage_calories_removed_tailored <- percentage_calories_removed_tailored %>%
  left_join(nonheto_densities_tailored) %>%
  mutate(across(c(heto_calories_removed, heto_nonheto_calories_remaining, nonheto_calories, nonheto_calories_per_ha), ~round(., digits = 0)))
```

## Step 13: Explore data

What plant species provided the most calories in each replicate?
```{r step13, fig.width=6, fig.height=4}
#Sizes for plot elements
size = 12
label_size = size
title_size = size * 1.15

#Assign abundance rank at the level of plant species, based on calories
calorie_abundance_ranks <- resource_data_to_summarize %>% 
  filter(count_estimate_high_low == "high") %>%
  group_by(count_estimate_high_low, year, patch, control_treatment, plant_species, plant_scientific_name) %>%
  summarise(num_calories = sum(calories_per_plant, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(count_estimate_high_low, year, patch, control_treatment) %>% 
  arrange(count_estimate_high_low, year, patch, control_treatment, desc(num_calories)) %>%
  mutate(rank = row_number()) %>%
  ungroup()

#In which replicates is HETO the most abundant source of calories?
heto_dominant_calorie_source <- calorie_abundance_ranks %>% 
  filter(rank == 1 & plant_species == "HETO")

#In which *treatment* replicates is HETO the most abundant source of calories?
heto_dominant_calorie_source_treatment <- calorie_abundance_ranks %>% 
  filter(rank == 1 & plant_species == "HETO" & control_treatment == "treatment")

#PLOT: what does the distribution of ranks look like for HETO? I.e., how many times is HETO ranked #1, #2, etc.
heto_rank_barplot <- calorie_abundance_ranks %>%
  filter(plant_species == "HETO") %>%
  group_by(rank) %>%
  summarise(count = n()) %>%
  ungroup() %>%
  ggplot(., aes(x = factor(rank), y = count)) + 
    geom_bar(stat = "identity", width = c(0.6), fill = "#BA0022") + 
    labs(x = "Rank", y = "Number of replicates", title = expression("Distribution of calorie abundance ranks for"~italic(Heliconia)~italic(tortuosa))) +
    theme_bw(base_size = size) + 
    theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank())

heto_rank_barplot
# ggview::ggview(heto_rank_barplot, width = 6, height = 4)

#PLOT: what species are usually the most abundant source of calories? I.e., for all the species that are ranked #1, how does HETO compare?
top_ranked_species_barplot <- calorie_abundance_ranks %>%
  filter(rank == 1) %>%
  group_by(plant_species, plant_scientific_name) %>%
  summarise(count = n()) %>%
  ungroup() %>%
  mutate(heto_color = ifelse(plant_species == "HETO", "yes", "no")) %>%
  make_barplot(xvar = "plant_scientific_name", yvar = "count", xlabel = "Plant species", ylabel = "Number of replicates where species was dominant calorie source", titlelabel = "Top-ranked plant species", highlight_column = "heto_color", text_size = size, flip_axes = TRUE)

top_ranked_species_barplot
# ggview::ggview(top_ranked_species_barplot, width = 6, height = 4)
```

What plant species were present in each replicate/site?

Need to communicate that *Heliconia tortuosa* really was the most common understory plant visited by hummingbirds! Even though other plant species were visited more on an hourly basis when they were present, only HETO and MARA were present in every replicate.
```{r}
#Occurrence per replicate (replicate = site-year combination)
plant_occurrence_per_replicate <- resource_data_to_summarize %>%
  distinct(year, patch, plant_family, plant_scientific_name, plant_species) %>%
  group_by(plant_family, plant_scientific_name, plant_species) %>%
  summarise(num_replicates = n()) %>%
  arrange(desc(num_replicates))

head(plant_occurrence_per_replicate)

#Occurrence per site
plant_occurrence_per_site <- resource_data_to_summarize %>%
  distinct(patch, plant_family, plant_scientific_name, plant_species) %>%
  group_by(plant_family, plant_scientific_name, plant_species) %>%
  summarise(num_sites = n()) %>%
  arrange(desc(num_sites))

head(plant_occurrence_per_site)
```

## Step 14: Export data
```{r step14}
#Data used in further analysis (i.e., data that will go onto Dryad)

#Combine data so that file can also be used in control vs. treatment comparison
site_and_replicate_characteristics <- heto_removed_as_treatment %>%
  left_join(heto_densities) %>% #for control vs. treatment comparison
  left_join(elevation) %>% #for control vs. treatment comparison
  select(year, patch, control_treatment, elevation, focal_area_size = area, everything())
  
write.csv(site_and_replicate_characteristics, "../../data/export/for_analysis/Site_and_replicate_characteristics.csv", row.names = FALSE)
write.csv(weights, "../../data/export/for_analysis/Weights.csv", row.names = FALSE)
write.csv(percentage_calories_removed_tailored, "../../data/export/for_analysis/Percentage_calories_removed.csv", row.names = FALSE)

#Data for Supporting information
write.csv(heto_removed_supp, "../../data/export/supp_tables/Heliconia_resources_removed.csv", row.names = FALSE)
write.csv(heto_removed_sum, "../../data/export/supp_tables/Heliconia_resources_removed_sum.csv", row.names = FALSE)

#extra tables - not including in final Supporting information
write.csv(weights_supp, "../../data/export/supp_tables/extra/Weights_supp.csv", row.names = FALSE)
write.csv(percentage_calories_removed_tailored_supp, "../../data/export/supp_tables/extra/Percentage_calories_removed_supp.csv", row.names = FALSE)
write.csv(percentage_calories_removed_tailored_sum, "../../data/export/supp_tables/extra/Percentage_calories_removed_sum.csv", row.names = FALSE)
write.csv(heto_densities_sum, "../../data/export/supp_tables/extra/Heliconia_densities_sum.csv", row.names = FALSE)
write.csv(species_per_site_sum, "../../data/export/supp_tables/extra/Plant_species_per_site_sum.csv", row.names = FALSE)
```
